requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
twilio
apscheduler
pytz
huggingface-hub
requests
